Add report

+++ app/controller/account/delete.php
@@ -7,22 +7,23 @@
     include_once('../../config/Database.php');
     include_once('../../models/Account.php');
 
+    if (isset($_SESSION['role']) && isset($_SESSION['isLogin']) && $_SESSION['role'] == 'Admin' && $_SESSION['isLogin'] == true) {
+        // database and connect
+        $account = new Account((new Database())->connect());
+        
+        // input data
+        $account->id = isset($_GET['id']) ? $_GET['id'] : die();
 
-    // database and connect
-    $account = new Account((new Database())->connect());
-    
-    // input data
-    $account->id = isset($_GET['id']) ? $_GET['id'] : die();
+        $result = $account->delete();
 
-    $result = $account->delete();
-
-    if ($result)
-        echo json_encode(
-            array('message' => 'success')
-        );
-    else {
-        echo json_encode(
-            array('message' => 'fail')
-        );
+        if ($result)
+            echo json_encode(
+                array('message' => 'success')
+            );
+        else {
+            echo json_encode(
+                array('message' => 'fail')
+            );
+        }
     }
 ?>
